requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image==0.16.2
tensorflow==1.15.0
tensorflow-gpu==1.15.0
keras==2.3.1
protobuf==3.20.1
pathlib
opencv-python
h5py==2.10.0
imgaug
IPython[all]
mrcnn
pika
opencv-python
pytesseract